Move tests per CR
